requirements.txt
Django==1.7
Padding==0.4
SQLAlchemy==0.9.8
dj-database-url==0.3.0
django-keen==0.1.3
django-postgrespool==0.3.0
djangorestframework==3.0.0
gunicorn==19.1.1
keen==0.3.0
psycopg2==2.5.4
pycrypto==2.6.1
requests==2.5.0
whitenoise==1.0.3
wsgiref==0.1.2
